+++ tests/testthat/test-cross_fit.R
@@ -213,7 +213,8 @@ test_that("invalid tidiers", {
 })
 
 test_that("abort if not formulas", {
-  expect_error(cross_fit(df, "x", m), "x.*is of type.*character")
-  expect_error(cross_fit(df, list("x"), m), "x.*is of type.*character")
-  expect_error(cross_fit(df, as.list(letters), m), "c.*is of type.*character")
+  expect_error(cross_fit(df, "x", m), "x.*is of class.*character")
+  expect_error(cross_fit(df, list("x"), m), "x.*is of class.*character")
+  expect_error(cross_fit(df, as.list(letters), m), "c.*is of class.*character")
+  expect_error(cross_fit(df, as.list(letters), m), "... and 21 more")
 })