+++ NEWS.md
@@ -1,3 +1,5 @@
+# crossmap (development version)
+
 # crossmap 0.3.0
 
 ## New features